#2338 trendingPartsView(date), prospect bug fixed.

+++ src/app/admin/prospects/prospects-edit/prospects-edit.component.ts
@@ -3,7 +3,9 @@ import { FormGroup, FormBuilder } from '@angular/forms';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
 import { Subscription } from 'rxjs';
 import { skipWhile } from 'rxjs/operators';
+import { LoginService } from 'src/app/services/login.service';
 import { AdministrativeService } from 'src/app/store/administrative-state-management/service/administrative.service';
+import { AppUtility } from 'src/app/utility/app.utility';
 import { SubscriptionUtil } from 'src/app/utility/subscription-utility';
 
 @Component({
@@ -18,10 +20,12 @@ export class ProspectsEditComponent implements OnInit, OnDestroy {
   private readonly subscriptions: Subscription = new Subscription();
   isForce = false;
   userId: any;
+  errorMessage: String;
   constructor(
     private readonly formBuilder: FormBuilder,
     private readonly administrativeService: AdministrativeService,
     private readonly el: ElementRef,
+    private readonly loginService: LoginService,
     public dialogRef: MatDialogRef<ProspectsEditComponent>,
     @Inject(MAT_DIALOG_DATA) public data: any) {
 
@@ -31,6 +35,7 @@ export class ProspectsEditComponent implements OnInit, OnDestroy {
     this.setForm(undefined);
     if (this.data !== undefined && this.data.id !== undefined) {
         this.id = this.data.id;
+      this.getProspectsById();
       this.loadProspectsById();
     }
   }
@@ -64,16 +69,26 @@ export class ProspectsEditComponent implements OnInit, OnDestroy {
 
   loadProspectsById() {
     this.administrativeService.loadProspectsById(this.data.id);
+  }
+
+  getProspectsById(){
     this.subscriptions.add(this.administrativeService.getProspectsById().pipe(skipWhile((item: any) => !item))
       .subscribe((prospects: any) => {
         this.setForm(prospects);
       }));
   }
+
   onNoClick() {
     this.dialogRef.close(false);
   }
 
   delete() {
+    this.errorMessage = '';
+
+    if(!confirm('Are you sure you want to delete?')){
+      return;
+    }
+
     this.subscriptions.add(this.administrativeService.deleteProspectsById(this.id).pipe(skipWhile((item: any) => !item))
       .subscribe((response: any) => {
         this.dialogRef.close(true);
@@ -81,19 +96,12 @@ export class ProspectsEditComponent implements OnInit, OnDestroy {
   }
 
   save() {
+    this.errorMessage = '';
     if (this.prospectsForm.valid) {
       if (this.id !== null && this.id !== undefined) {
-        this.subscriptions.add(this.administrativeService.updateProspects(this.id, this.prospectsForm.value).pipe(
-          skipWhile((item: any) => !item))
-          .subscribe((response: any) => {
-            this.loadProspectsById();
-          }));
+        this.administrativeService.updateProspects(this.id, this.prospectsForm.value);
       } else {
-        this.subscriptions.add(this.administrativeService.saveProspects(this.prospectsForm.value).pipe(
-          skipWhile((item: any) => !item))
-          .subscribe((response: any) => {
-            this.loadProspectsById();
-          }));
+        this.administrativeService.saveProspects(this.prospectsForm.value);
       }
       this.dialogRef.close(true);
     } else {
@@ -108,6 +116,26 @@ export class ProspectsEditComponent implements OnInit, OnDestroy {
         break;
       }
     }
+  }
+
+  resendEmailValidation(){
+    this.loginService.performPost({},'/trusted/registrations/' + this.data.id +'/sendValidationMail')
+    .subscribe(
+      response => {
+        console.log(response);
+
+        if(response.data.errorMessage){
+          this.errorMessage = response.data.errorMessage;
+          console.error(response.data);
+          console.log(this.el);
+          document.querySelector('mat-dialog-container').scrollTo(0,0);
+        }
+
+      }, error =>{
+        console.error(error);
+        AppUtility.scrollTop();
+      }
+    )
   }
 
   get f() { return this.prospectsForm.controls; }
